\name{tProfile}
\alias{tProfile}
\title{Temperature Profile Data of Bautzen Reservoir}

\description{The dataset is a one year subset of water temperature data, 
 measured during 1995 in a shallow reservoir.
}

\format{tProfile is a data frame with the first column giving the Julian 
        day of the measurement and 11 columns giving the water temperature of 
        each depth from surface (0 m) up to 10 m below surface.}

\usage{data(tProfile)}

\source{Dresden University of Technology, Institute of Hydrobiology}

\keyword{datasets}
